Split ORM models into per-subdomain orm/ modules (#421)
ORM models lived in two contradictory places: a 760-line src/models.py
monolith and a lone infrastructure/jobs/orm/ package. Standardise on the
jobs convention by moving every model into its owning subdomain's orm/
package, mirroring the domain/<sub>/entities/ layout.

- identity/orm: User, RefreshToken
- library/orm: Book, Chapter
- reading/orm: Highlight, HighlightStyle, Tag, TagGroup, Bookmark,
  ReadingSession, ChapterPrereadingContent + associations
- learning/orm: Flashcard, AIChatSession
- ai/orm: AIUsageRecord
- notes/orm: Note + associations

src/models.py becomes a thin aggregator that imports every ORM module so
Base.metadata stays complete for Alembic autogenerate and create_all, and
re-exports the classes for convenience. Repositories and mappers now import
from their subdomain orm/ package directly, which also breaks the import
cycle between src.models and the DI container. Cross-subdomain relationships
use TYPE_CHECKING imports with SQLAlchemy string forward-refs, so there are
no runtime import cycles.

Verified: pyright clean, ruff clean, 433 tests pass, metadata table set
identical (22 tables, no schema drift).
