+++ backend/src/infrastructure/learning/repositories/ai_chat_session_repository.py
@@ -7,7 +7,9 @@ from src.domain.common.types import SerializedMessageHistory
 from src.domain.common.value_objects.ids import AIChatSessionId, UserId
 from src.domain.learning.entities.ai_chat_session import AIChatSession
 from src.infrastructure.learning.mappers.ai_chat_session_mapper import AIChatSessionMapper
-from src.models import AIChatSession as AIChatSessionModel
+from src.infrastructure.learning.orm.ai_chat_session_model import (
+    AIChatSession as AIChatSessionModel,
+)
 
 
 class AIChatSessionRepository:
