+++ backend/src/infrastructure/reading/mappers/highlight_style_mapper.py
@@ -3,7 +3,7 @@
 from src.domain.common.value_objects import BookId, HighlightStyleId, UserId
 from src.domain.reading.entities.highlight_style import HighlightStyle
 from src.infrastructure.common.mappers import orm_id
-from src.models import HighlightStyle as HighlightStyleORM
+from src.infrastructure.reading.orm.highlight_style_model import HighlightStyle as HighlightStyleORM
 
 
 class HighlightStyleMapper:
