+++ backend/src/models.py
@@ -1,760 +1,64 @@
-"""Database models."""
+"""ORM model aggregator.
 
-from datetime import datetime as dt
-from typing import Any
-
-from sqlalchemy import (
-    Column,
-    DateTime,
-    ForeignKey,
-    Index,
-    Integer,
-    String,
-    Table,
-    Text,
-    UniqueConstraint,
-    func,
-)
-from sqlalchemy.dialects.postgresql import JSONB as PG_JSONB
-from sqlalchemy.dialects.postgresql import TSVECTOR
-from sqlalchemy.orm import Mapped, mapped_column, relationship
-from sqlalchemy.types import JSON
+ORM models live per-subdomain under ``src/infrastructure/<subdomain>/orm/``.
+This module imports every one of them so that they are registered on
+``Base.metadata`` (used by Alembic autogenerate and table creation) and
+re-exports them for convenience. Prefer importing models from their owning
+subdomain's ``orm`` package directly.
+"""
 
 from src.database import Base
+from src.infrastructure.ai.orm.ai_usage_record_model import AIUsageRecord
+from src.infrastructure.identity.orm.refresh_token_model import RefreshToken
+from src.infrastructure.identity.orm.user_model import User
+from src.infrastructure.jobs.orm.job_batch_model import JobBatchModel
+from src.infrastructure.learning.orm.ai_chat_session_model import AIChatSession
+from src.infrastructure.learning.orm.flashcard_model import Flashcard
+from src.infrastructure.library.orm.book_model import Book
+from src.infrastructure.library.orm.chapter_model import Chapter
+from src.infrastructure.notes.orm.associations import (
+    note_books,
+    note_chapters,
+    note_highlights,
+    note_tags,
+)
+from src.infrastructure.notes.orm.note_model import Note
+from src.infrastructure.reading.orm.associations import (
+    highlight_tags,
+    reading_session_highlights,
+)
+from src.infrastructure.reading.orm.bookmark_model import Bookmark
+from src.infrastructure.reading.orm.chapter_prereading_content_model import (
+    ChapterPrereadingContent,
+)
+from src.infrastructure.reading.orm.highlight_model import Highlight
+from src.infrastructure.reading.orm.highlight_style_model import HighlightStyle
+from src.infrastructure.reading.orm.reading_session_model import ReadingSession
+from src.infrastructure.reading.orm.tag_group_model import TagGroup
+from src.infrastructure.reading.orm.tag_model import Tag
 
-# Association table for many-to-many relationship between highlights and tags
-highlight_tags = Table(
+__all__ = [
+    "AIChatSession",
+    "AIUsageRecord",
+    "Base",
+    "Book",
+    "Bookmark",
+    "Chapter",
+    "ChapterPrereadingContent",
+    "Flashcard",
+    "Highlight",
+    "HighlightStyle",
+    "JobBatchModel",
+    "Note",
+    "ReadingSession",
+    "RefreshToken",
+    "Tag",
+    "TagGroup",
+    "User",
     "highlight_tags",
-    Base.metadata,
-    Column(
-        "highlight_id",
-        Integer,
-        ForeignKey("highlights.id", ondelete="CASCADE"),
-        primary_key=True,
-        index=True,
-    ),
-    Column(
-        "tag_id",
-        Integer,
-        ForeignKey("tags.id", ondelete="CASCADE"),
-        primary_key=True,
-        index=True,
-    ),
-    Column(
-        "created_at",
-        DateTime(timezone=True),
-        server_default=func.now(),
-        nullable=False,
-    ),
-)
-
-
-# Association table for many-to-many relationship between reading_sessions and highlights
-reading_session_highlights = Table(
-    "reading_session_highlights",
-    Base.metadata,
-    Column(
-        "reading_session_id",
-        Integer,
-        ForeignKey("reading_sessions.id", ondelete="CASCADE"),
-        primary_key=True,
-    ),
-    Column(
-        "highlight_id",
-        Integer,
-        ForeignKey("highlights.id", ondelete="CASCADE"),
-        primary_key=True,
-        index=True,
-    ),
-    Column(
-        "created_at",
-        DateTime(timezone=True),
-        server_default=func.now(),
-        nullable=False,
-    ),
-)
-
-
-# Association tables for many-to-many relationships between notes and other entities
-note_books = Table(
     "note_books",
-    Base.metadata,
-    Column(
-        "note_id", Integer, ForeignKey("notes.id", ondelete="CASCADE"), primary_key=True, index=True
-    ),
-    Column(
-        "book_id", Integer, ForeignKey("books.id", ondelete="CASCADE"), primary_key=True, index=True
-    ),
-    Column(
-        "created_at",
-        DateTime(timezone=True),
-        server_default=func.now(),
-        nullable=False,
-    ),
-)
-
-note_chapters = Table(
     "note_chapters",
-    Base.metadata,
-    Column(
-        "note_id", Integer, ForeignKey("notes.id", ondelete="CASCADE"), primary_key=True, index=True
-    ),
-    Column(
-        "chapter_id",
-        Integer,
-        ForeignKey("chapters.id", ondelete="CASCADE"),
-        primary_key=True,
-        index=True,
-    ),
-    Column(
-        "created_at",
-        DateTime(timezone=True),
-        server_default=func.now(),
-        nullable=False,
-    ),
-)
-
-note_highlights = Table(
     "note_highlights",
-    Base.metadata,
-    Column(
-        "note_id", Integer, ForeignKey("notes.id", ondelete="CASCADE"), primary_key=True, index=True
-    ),
-    Column(
-        "highlight_id",
-        Integer,
-        ForeignKey("highlights.id", ondelete="CASCADE"),
-        primary_key=True,
-        index=True,
-    ),
-    Column(
-        "created_at",
-        DateTime(timezone=True),
-        server_default=func.now(),
-        nullable=False,
-    ),
-)
-
-note_tags = Table(
     "note_tags",
-    Base.metadata,
-    Column(
-        "note_id", Integer, ForeignKey("notes.id", ondelete="CASCADE"), primary_key=True, index=True
-    ),
-    Column(
-        "tag_id",
-        Integer,
-        ForeignKey("tags.id", ondelete="CASCADE"),
-        primary_key=True,
-        index=True,
-    ),
-    Column(
-        "created_at",
-        DateTime(timezone=True),
-        server_default=func.now(),
-        nullable=False,
-    ),
-)
-
-
-class User(Base):
-    """User model for multi-user support."""
-
-    __tablename__ = "users"
-
-    id: Mapped[int] = mapped_column(primary_key=True, index=True)
-    email: Mapped[str] = mapped_column(String(100), nullable=False, unique=True)
-    hashed_password: Mapped[str | None] = mapped_column(String(255), nullable=True)
-    created_at: Mapped[dt] = mapped_column(
-        DateTime(timezone=True), server_default=func.now(), nullable=False
-    )
-    updated_at: Mapped[dt] = mapped_column(
-        DateTime(timezone=True),
-        server_default=func.now(),
-        onupdate=func.now(),
-        nullable=False,
-    )
-
-    # Relationships
-    books: Mapped[list["Book"]] = relationship(back_populates="user", cascade="all, delete-orphan")
-    highlights: Mapped[list["Highlight"]] = relationship(
-        back_populates="user", cascade="all, delete-orphan"
-    )
-    highlight_styles: Mapped[list["HighlightStyle"]] = relationship(
-        back_populates="user", cascade="all, delete-orphan"
-    )
-    tags: Mapped[list["Tag"]] = relationship(back_populates="user", cascade="all, delete-orphan")
-    flashcards: Mapped[list["Flashcard"]] = relationship(
-        back_populates="user", cascade="all, delete-orphan"
-    )
-    reading_sessions: Mapped[list["ReadingSession"]] = relationship(
-        back_populates="user", cascade="all, delete-orphan"
-    )
-    refresh_tokens: Mapped[list["RefreshToken"]] = relationship(
-        back_populates="user", cascade="all, delete-orphan"
-    )
-
-    def __repr__(self) -> str:
-        """String representation of User."""
-        return f"<User(id={self.id}, email='{self.email}')>"
-
-
-class RefreshToken(Base):
-    """Refresh token for token rotation and replay detection."""
-
-    __tablename__ = "refresh_tokens"
-
-    id: Mapped[int] = mapped_column(primary_key=True, index=True)
-    jti: Mapped[str] = mapped_column(String(36), nullable=False, unique=True, index=True)
-    user_id: Mapped[int] = mapped_column(
-        ForeignKey("users.id", ondelete="CASCADE"), nullable=False, index=True
-    )
-    family_id: Mapped[str] = mapped_column(String(36), nullable=False, index=True)
-    revoked_at: Mapped[dt | None] = mapped_column(DateTime(timezone=True), nullable=True)
-    expires_at: Mapped[dt] = mapped_column(DateTime(timezone=True), nullable=False)
-    created_at: Mapped[dt] = mapped_column(
-        DateTime(timezone=True), server_default=func.now(), nullable=False
-    )
-
-    user: Mapped["User"] = relationship(back_populates="refresh_tokens")
-
-    def __repr__(self) -> str:
-        return f"<RefreshToken(id={self.id}, jti='{self.jti}', user_id={self.user_id})>"
-
-
-class Book(Base):
-    """Book model for storing book metadata."""
-
-    __tablename__ = "books"
-
-    id: Mapped[int] = mapped_column(primary_key=True, index=True)
-    user_id: Mapped[int] = mapped_column(
-        ForeignKey("users.id", ondelete="CASCADE"), index=True, nullable=False
-    )
-    title: Mapped[str] = mapped_column(String(500), nullable=False)
-    author: Mapped[str | None] = mapped_column(String(500), nullable=True)
-    isbn: Mapped[str | None] = mapped_column(String(20), nullable=True)
-    ebook_file: Mapped[str | None] = mapped_column(String(500), nullable=True)
-    file_type: Mapped[str | None] = mapped_column(String(10), nullable=True)
-    cover_file: Mapped[str | None] = mapped_column(String(500), nullable=True)
-    cover_blurhash: Mapped[str | None] = mapped_column(String(40), nullable=True)
-    description: Mapped[str | None] = mapped_column(Text, nullable=True)
-    language: Mapped[str | None] = mapped_column(String(10), nullable=True)
-    page_count: Mapped[int | None] = mapped_column(nullable=True)
-    client_book_id: Mapped[str | None] = mapped_column(
-        String(255), nullable=True, index=True
-    )  # Client-provided stable book identifier
-    created_at: Mapped[dt] = mapped_column(
-        DateTime(timezone=True), server_default=func.now(), nullable=False
-    )
-    updated_at: Mapped[dt] = mapped_column(
-        DateTime(timezone=True),
-        server_default=func.now(),
-        onupdate=func.now(),
-        nullable=False,
-    )
-    last_viewed: Mapped[dt | None] = mapped_column(
-        DateTime(timezone=True), nullable=True, index=True
-    )
-    end_position: Mapped[list[int] | None] = mapped_column(JSON, nullable=True)
-
-    # Relationships
-    user: Mapped["User"] = relationship(back_populates="books")
-    highlights: Mapped[list["Highlight"]] = relationship(
-        back_populates="book", cascade="all, delete-orphan"
-    )
-    chapters: Mapped[list["Chapter"]] = relationship(
-        back_populates="book", cascade="all, delete-orphan"
-    )
-    tags: Mapped[list["Tag"]] = relationship(back_populates="book", cascade="all, delete-orphan")
-    highlight_styles: Mapped[list["HighlightStyle"]] = relationship(
-        back_populates="book", cascade="all, delete-orphan"
-    )
-    tag_groups: Mapped[list["TagGroup"]] = relationship(
-        back_populates="book", cascade="all, delete-orphan", lazy="selectin"
-    )
-    bookmarks: Mapped[list["Bookmark"]] = relationship(
-        back_populates="book", cascade="all, delete-orphan"
-    )
-    flashcards: Mapped[list["Flashcard"]] = relationship(
-        back_populates="book", cascade="all, delete-orphan"
-    )
-    reading_sessions: Mapped[list["ReadingSession"]] = relationship(
-        back_populates="book", cascade="all, delete-orphan"
-    )
-
-    def __repr__(self) -> str:
-        """String representation of Book."""
-        return f"<Book(id={self.id}, title='{self.title}')>"
-
-
-class Chapter(Base):
-    """Chapter model for storing unique chapters within a book."""
-
-    __tablename__ = "chapters"
-
-    id: Mapped[int] = mapped_column(primary_key=True, index=True)
-    book_id: Mapped[int] = mapped_column(
-        ForeignKey("books.id", ondelete="CASCADE"), index=True, nullable=False
-    )
-    parent_id: Mapped[int | None] = mapped_column(
-        ForeignKey("chapters.id", ondelete="CASCADE"), index=True, nullable=True
-    )
-    name: Mapped[str] = mapped_column(String(500), nullable=False)
-    chapter_number: Mapped[int | None] = mapped_column(nullable=True, index=True)
-    start_xpoint: Mapped[str | None] = mapped_column(Text, nullable=True)
-    end_xpoint: Mapped[str | None] = mapped_column(Text, nullable=True)
-    start_position: Mapped[list[int] | None] = mapped_column(JSON, nullable=True)
-    end_position: Mapped[list[int] | None] = mapped_column(JSON, nullable=True)
-    created_at: Mapped[dt] = mapped_column(
-        DateTime(timezone=True), server_default=func.now(), nullable=False
-    )
-    updated_at: Mapped[dt] = mapped_column(
-        DateTime(timezone=True),
-        server_default=func.now(),
-        onupdate=func.now(),
-        nullable=False,
-    )
-
-    # Relationships
-    book: Mapped["Book"] = relationship(back_populates="chapters")
-    highlights: Mapped[list["Highlight"]] = relationship(back_populates="chapter")
-    prereading_content: Mapped["ChapterPrereadingContent | None"] = relationship(
-        back_populates="chapter", uselist=False, cascade="all, delete-orphan"
-    )
-    # Self-referential relationship for hierarchical chapters
-    parent: Mapped["Chapter | None"] = relationship(
-        "Chapter", remote_side="Chapter.id", back_populates="children"
-    )
-    children: Mapped[list["Chapter"]] = relationship(
-        "Chapter", back_populates="parent", cascade="all, delete-orphan"
-    )
-
-    # Unique constraint: chapter names must be unique within same parent level
-    # (allows "Harjoitukset" under Part I and "Harjoitukset" under Part II)
-    __table_args__ = (UniqueConstraint("book_id", "parent_id", "name", name="uq_chapter_per_book"),)
-
-    def __repr__(self) -> str:
-        """String representation of Chapter."""
-        return f"<Chapter(id={self.id}, name='{self.name}', book_id={self.book_id})>"
-
-
-class HighlightStyle(Base):
-    """HighlightStyle model for storing highlight style labels and colors."""
-
-    __tablename__ = "highlight_styles"
-
-    id: Mapped[int] = mapped_column(primary_key=True, index=True)
-    user_id: Mapped[int] = mapped_column(
-        ForeignKey("users.id", ondelete="CASCADE"), index=True, nullable=False
-    )
-    book_id: Mapped[int | None] = mapped_column(
-        ForeignKey("books.id", ondelete="CASCADE"), index=True, nullable=True
-    )
-    device_color: Mapped[str | None] = mapped_column(String(50), nullable=True)
-    device_style: Mapped[str | None] = mapped_column(String(50), nullable=True)
-    label: Mapped[str | None] = mapped_column(String(255), nullable=True)
-    ui_color: Mapped[str | None] = mapped_column(String(20), nullable=True)
-    created_at: Mapped[dt] = mapped_column(
-        DateTime(timezone=True), server_default=func.now(), nullable=False
-    )
-    updated_at: Mapped[dt] = mapped_column(
-        DateTime(timezone=True),
-        server_default=func.now(),
-        onupdate=func.now(),
-        nullable=False,
-    )
-
-    # Relationships
-    user: Mapped["User"] = relationship(back_populates="highlight_styles")
-    book: Mapped["Book | None"] = relationship(back_populates="highlight_styles")
-    highlights: Mapped[list["Highlight"]] = relationship(back_populates="highlight_style_rel")
-
-    def __repr__(self) -> str:
-        """String representation of HighlightStyle."""
-        return (
-            f"<HighlightStyle(id={self.id}, user_id={self.user_id}, "
-            f"device_color='{self.device_color}', device_style='{self.device_style}')>"
-        )
-
-
-class Highlight(Base):
-    """Highlight model for storing KOReader highlights."""
-
-    __tablename__ = "highlights"
-
-    id: Mapped[int] = mapped_column(primary_key=True, index=True)
-    user_id: Mapped[int] = mapped_column(
-        ForeignKey("users.id", ondelete="CASCADE"), index=True, nullable=False
-    )
-    book_id: Mapped[int] = mapped_column(
-        ForeignKey("books.id", ondelete="CASCADE"), index=True, nullable=False
-    )
-    chapter_id: Mapped[int | None] = mapped_column(
-        ForeignKey("chapters.id", ondelete="SET NULL"), index=True, nullable=True
-    )
-    text: Mapped[str] = mapped_column(Text, nullable=False)
-    page: Mapped[int | None] = mapped_column(nullable=True)
-    start_xpoint: Mapped[str | None] = mapped_column(Text, nullable=True)
-    end_xpoint: Mapped[str | None] = mapped_column(Text, nullable=True)
-    position: Mapped[list[int] | None] = mapped_column(JSON, nullable=True)
-    highlight_style_id: Mapped[int | None] = mapped_column(
-        ForeignKey("highlight_styles.id", ondelete="SET NULL"), index=True, nullable=True
-    )
-    datetime: Mapped[str] = mapped_column(String(50), nullable=False)  # KOReader datetime string
-    content_hash: Mapped[str] = mapped_column(
-        String(64), nullable=False, index=True
-    )  # SHA-256 hash for deduplication
-    created_at: Mapped[dt] = mapped_column(
-        DateTime(timezone=True), server_default=func.now(), nullable=False
-    )
-    updated_at: Mapped[dt] = mapped_column(
-        DateTime(timezone=True),
-        server_default=func.now(),
-        onupdate=func.now(),
-        nullable=False,
-    )
-    deleted_at: Mapped[dt | None] = mapped_column(
-        DateTime(timezone=True), nullable=True, index=True
-    )
-    text_search_vector: Mapped[str | None] = mapped_column(
-        Text().with_variant(TSVECTOR, "postgresql"), nullable=True, index=True
-    )
-
-    # Relationships
-    user: Mapped["User"] = relationship(back_populates="highlights")
-    book: Mapped["Book"] = relationship(back_populates="highlights")
-    chapter: Mapped["Chapter | None"] = relationship(back_populates="highlights")
-    highlight_style_rel: Mapped["HighlightStyle | None"] = relationship(back_populates="highlights")
-    tags: Mapped[list["Tag"]] = relationship(
-        secondary=highlight_tags, back_populates="highlights", lazy="selectin"
-    )
-    bookmarks: Mapped[list["Bookmark"]] = relationship(
-        back_populates="highlight", cascade="all, delete-orphan"
-    )
-    flashcards: Mapped[list["Flashcard"]] = relationship(
-        back_populates="highlight", cascade="all, delete-orphan"
-    )
-    reading_sessions: Mapped[list["ReadingSession"]] = relationship(
-        secondary=reading_session_highlights, back_populates="highlights", lazy="selectin"
-    )
-
-    # Unique constraint for deduplication: same content hash within same book
-    __table_args__ = (
-        UniqueConstraint("user_id", "book_id", "content_hash", name="uq_highlight_content_hash"),
-    )
-
-    def __repr__(self) -> str:
-        """String representation of Highlight."""
-        return f"<Highlight(id={self.id}, text='{self.text[:50]}...', book_id={self.book_id})>"
-
-
-class TagGroup(Base):
-    """TagGroup model for grouping tags within a book."""
-
-    __tablename__ = "tag_groups"
-
-    id: Mapped[int] = mapped_column(primary_key=True, index=True)
-    book_id: Mapped[int] = mapped_column(
-        ForeignKey("books.id", ondelete="CASCADE"), index=True, nullable=False
-    )
-    name: Mapped[str] = mapped_column(String(100), nullable=False, index=True)
-    created_at: Mapped[dt] = mapped_column(
-        DateTime(timezone=True), server_default=func.now(), nullable=False
-    )
-    updated_at: Mapped[dt] = mapped_column(
-        DateTime(timezone=True),
-        server_default=func.now(),
-        onupdate=func.now(),
-        nullable=False,
-    )
-
-    # Relationships
-    book: Mapped["Book"] = relationship(back_populates="tag_groups")
-    tags: Mapped[list["Tag"]] = relationship(back_populates="tag_group")
-
-    # Unique constraint: tag group names are unique within a book
-    __table_args__ = (UniqueConstraint("book_id", "name", name="uq_tag_group_book_name"),)
-
-    def __repr__(self) -> str:
-        """String representation of TagGroup."""
-        return f"<TagGroup(id={self.id}, name='{self.name}', book_id={self.book_id})>"
-
-
-class Tag(Base):
-    """Tag model for categorizing highlights within a book."""
-
-    __tablename__ = "tags"
-
-    id: Mapped[int] = mapped_column(primary_key=True, index=True)
-    user_id: Mapped[int] = mapped_column(
-        ForeignKey("users.id", ondelete="CASCADE"), index=True, nullable=False
-    )
-    book_id: Mapped[int] = mapped_column(
-        ForeignKey("books.id", ondelete="CASCADE"), index=True, nullable=False
-    )
-    tag_group_id: Mapped[int | None] = mapped_column(
-        ForeignKey("tag_groups.id", ondelete="SET NULL"), index=True, nullable=True
-    )
-    name: Mapped[str] = mapped_column(String(100), nullable=False, index=True)
-    created_at: Mapped[dt] = mapped_column(
-        DateTime(timezone=True), server_default=func.now(), nullable=False
-    )
-    updated_at: Mapped[dt] = mapped_column(
-        DateTime(timezone=True),
-        server_default=func.now(),
-        onupdate=func.now(),
-        nullable=False,
-    )
-
-    # Relationships
-    user: Mapped["User"] = relationship(back_populates="tags")
-    book: Mapped["Book"] = relationship(back_populates="tags")
-    tag_group: Mapped["TagGroup | None"] = relationship(back_populates="tags", lazy="selectin")
-    highlights: Mapped[list["Highlight"]] = relationship(
-        secondary=highlight_tags, back_populates="tags", lazy="selectin"
-    )
-
-    # Unique constraint: tag names are unique within a book per user
-    __table_args__ = (UniqueConstraint("user_id", "book_id", "name", name="uq_tag_user_book_name"),)
-
-    def __repr__(self) -> str:
-        """String representation of Tag."""
-        return f"<Tag(id={self.id}, name='{self.name}', book_id={self.book_id})>"
-
-
-class Bookmark(Base):
-    """Bookmark model for tracking reading progress within a book."""
-
-    __tablename__ = "bookmarks"
-
-    id: Mapped[int] = mapped_column(primary_key=True, index=True)
-    book_id: Mapped[int] = mapped_column(
-        ForeignKey("books.id", ondelete="CASCADE"), index=True, nullable=False
-    )
-    highlight_id: Mapped[int] = mapped_column(
-        ForeignKey("highlights.id", ondelete="CASCADE"), index=True, nullable=False
-    )
-    created_at: Mapped[dt] = mapped_column(
-        DateTime(timezone=True), server_default=func.now(), nullable=False
-    )
-
-    # Relationships
-    book: Mapped["Book"] = relationship(back_populates="bookmarks")
-    highlight: Mapped["Highlight"] = relationship(back_populates="bookmarks")
-
-    def __repr__(self) -> str:
-        """String representation of Bookmark."""
-        return f"<Bookmark(id={self.id}, book_id={self.book_id}, highlight_id={self.highlight_id})>"
-
-
-class Flashcard(Base):
-    """Flashcard model for creating study cards from highlights."""
-
-    __tablename__ = "flashcards"
-
-    id: Mapped[int] = mapped_column(primary_key=True, index=True)
-    user_id: Mapped[int] = mapped_column(
-        ForeignKey("users.id", ondelete="CASCADE"), index=True, nullable=False
-    )
-    book_id: Mapped[int] = mapped_column(
-        ForeignKey("books.id", ondelete="CASCADE"), index=True, nullable=False
-    )
-    highlight_id: Mapped[int | None] = mapped_column(
-        ForeignKey("highlights.id", ondelete="CASCADE"), index=True, nullable=True
-    )
-    chapter_id: Mapped[int | None] = mapped_column(
-        ForeignKey("chapters.id", ondelete="CASCADE"), index=True, nullable=True
-    )
-    # SET NULL: notes are hard-deleted; the flashcard survives as a book-level card
-    note_id: Mapped[int | None] = mapped_column(
-        ForeignKey("notes.id", ondelete="SET NULL"), index=True, nullable=True
-    )
-    question: Mapped[str] = mapped_column(Text, nullable=False)
-    answer: Mapped[str] = mapped_column(Text, nullable=False)
-    created_at: Mapped[dt] = mapped_column(
-        DateTime(timezone=True), server_default=func.now(), nullable=False
-    )
-    updated_at: Mapped[dt] = mapped_column(
-        DateTime(timezone=True),
-        server_default=func.now(),
-        onupdate=func.now(),
-        nullable=False,
-    )
-
-    # Relationships
-    user: Mapped["User"] = relationship(back_populates="flashcards")
-    book: Mapped["Book"] = relationship(back_populates="flashcards")
-    highlight: Mapped["Highlight | None"] = relationship(back_populates="flashcards")
-    chapter: Mapped["Chapter | None"] = relationship()
-
-    def __repr__(self) -> str:
-        """String representation of Flashcard."""
-        return (
-            f"<Flashcard(id={self.id}, book_id={self.book_id}, highlight_id={self.highlight_id})>"
-        )
-
-
-class ReadingSession(Base):
-    """ReadingSession model for tracking reading sessions from KOReader."""
-
-    __tablename__ = "reading_sessions"
-
-    id: Mapped[int] = mapped_column(primary_key=True, index=True)
-    user_id: Mapped[int] = mapped_column(
-        ForeignKey("users.id", ondelete="CASCADE"), index=True, nullable=False
-    )
-    book_id: Mapped[int] = mapped_column(
-        ForeignKey("books.id", ondelete="CASCADE"), index=True, nullable=False
-    )
-    start_time: Mapped[dt] = mapped_column(DateTime(timezone=True), nullable=False)
-    end_time: Mapped[dt] = mapped_column(DateTime(timezone=True), nullable=False)
-    start_xpoint: Mapped[str | None] = mapped_column(Text, nullable=True)
-    end_xpoint: Mapped[str | None] = mapped_column(Text, nullable=True)
-    start_page: Mapped[int | None] = mapped_column(nullable=True)
-    end_page: Mapped[int | None] = mapped_column(nullable=True)
-    start_position: Mapped[list[int] | None] = mapped_column(JSON, nullable=True)
-    end_position: Mapped[list[int] | None] = mapped_column(JSON, nullable=True)
-    device_id: Mapped[str | None] = mapped_column(String(100), nullable=True)
-    content_hash: Mapped[str] = mapped_column(
-        String(64), nullable=False, index=True
-    )  # SHA-256 hash for deduplication
-    created_at: Mapped[dt] = mapped_column(
-        DateTime(timezone=True), server_default=func.now(), nullable=False
-    )
-    ai_summary: Mapped[str | None] = mapped_column(Text, nullable=True)
-
-    # Relationships
-    user: Mapped["User"] = relationship(back_populates="reading_sessions")
-    book: Mapped["Book"] = relationship(back_populates="reading_sessions")
-    highlights: Mapped[list["Highlight"]] = relationship(
-        secondary=reading_session_highlights, back_populates="reading_sessions", lazy="selectin"
-    )
-
-    # Unique constraint for deduplication: same content hash for same user
-    __table_args__ = (
-        UniqueConstraint("user_id", "content_hash", name="uq_reading_session_content_hash"),
-    )
-
-    def __repr__(self) -> str:
-        """String representation of ReadingSession."""
-        return (
-            f"<ReadingSession(id={self.id}, book_id={self.book_id}, start_time={self.start_time})>"
-        )
-
-
-class ChapterPrereadingContent(Base):
-    """ORM model for chapter pre-reading content."""
-
-    __tablename__ = "chapter_prereading_contents"
-
-    id: Mapped[int] = mapped_column(primary_key=True, index=True)
-    chapter_id: Mapped[int] = mapped_column(
-        ForeignKey("chapters.id", ondelete="CASCADE"), nullable=False, unique=True, index=True
-    )
-    summary: Mapped[str] = mapped_column(Text, nullable=False)
-    keypoints: Mapped[list[str]] = mapped_column(
-        JSON().with_variant(PG_JSONB, "postgresql"), nullable=False
-    )
-    questions: Mapped[list[dict[str, str]]] = mapped_column(
-        JSON().with_variant(PG_JSONB, "postgresql"), nullable=False, server_default="[]"
-    )
-    generated_at: Mapped[dt] = mapped_column(DateTime(timezone=True), nullable=False)
-    ai_model: Mapped[str] = mapped_column(String(100), nullable=False)
-
-    # Relationships
-    chapter: Mapped["Chapter"] = relationship(back_populates="prereading_content")
-
-    def __repr__(self) -> str:
-        """String representation of ChapterPrereadingContent."""
-        return f"<ChapterPrereadingContent(id={self.id}, chapter_id={self.chapter_id})>"
-
-
-class AIUsageRecord(Base):
-    """ORM model for AI usage tracking."""
-
-    __tablename__ = "ai_usage_records"
-
-    id: Mapped[int] = mapped_column(primary_key=True, index=True)
-    user_id: Mapped[int] = mapped_column(
-        ForeignKey("users.id", ondelete="CASCADE"), nullable=False, index=True
-    )
-    task_type: Mapped[str] = mapped_column(String(50), nullable=False)
-    entity_type: Mapped[str] = mapped_column(String(50), nullable=False)
-    entity_id: Mapped[int] = mapped_column(nullable=False)
-    model_name: Mapped[str] = mapped_column(String(100), nullable=False)
-    input_tokens: Mapped[int] = mapped_column(nullable=False)
-    output_tokens: Mapped[int] = mapped_column(nullable=False)
-    created_at: Mapped[dt] = mapped_column(DateTime(timezone=True), nullable=False)
-
-    __table_args__ = (Index("ix_ai_usage_records_entity", "entity_type", "entity_id"),)
-
-    user: Mapped["User"] = relationship()
-
-    def __repr__(self) -> str:
-        return f"<AIUsageRecord(id={self.id}, task_type={self.task_type}, entity_type={self.entity_type})>"
-
-
-class AIChatSession(Base):
-    """ORM model for AI chat sessions (quiz, discussion, etc.)."""
-
-    __tablename__ = "ai_chat_sessions"
-
-    id: Mapped[int] = mapped_column(primary_key=True, index=True)
-    user_id: Mapped[int] = mapped_column(
-        ForeignKey("users.id", ondelete="CASCADE"), nullable=False, index=True
-    )
-    chapter_id: Mapped[int] = mapped_column(
-        ForeignKey("chapters.id", ondelete="CASCADE"), nullable=False, index=True
-    )
-    session_type: Mapped[str] = mapped_column(String(50), nullable=False, index=True)
-    message_history: Mapped[list[dict[str, Any]]] = mapped_column(JSON, nullable=False)
-    created_at: Mapped[dt] = mapped_column(
-        DateTime(timezone=True), server_default=func.now(), nullable=False
-    )
-
-    user: Mapped["User"] = relationship()
-    chapter: Mapped["Chapter"] = relationship()
-
-    def __repr__(self) -> str:
-        return (
-            f"<AIChatSession(id={self.id}, type={self.session_type}, chapter_id={self.chapter_id})>"
-        )
-
-
-from src.infrastructure.jobs.orm.job_batch_model import (  # noqa: E402
-    JobBatchModel,  # noqa: F401  # pyright: ignore[reportUnusedImport]
-)
-
-
-class Note(Base):
-    """User-authored note about terms, characters, or concepts in books."""
-
-    __tablename__ = "notes"
-
-    id: Mapped[int] = mapped_column(primary_key=True, index=True)
-    user_id: Mapped[int] = mapped_column(
-        ForeignKey("users.id", ondelete="CASCADE"), index=True, nullable=False
-    )
-    title: Mapped[str] = mapped_column(String(500), nullable=False)
-    body: Mapped[str] = mapped_column(Text, nullable=False, default="")
-    kind: Mapped[str | None] = mapped_column(String(20), nullable=True, index=True)
-    created_at: Mapped[dt] = mapped_column(
-        DateTime(timezone=True), server_default=func.now(), nullable=False
-    )
-    updated_at: Mapped[dt] = mapped_column(
-        DateTime(timezone=True),
-        server_default=func.now(),
-        onupdate=func.now(),
-        nullable=False,
-    )
-
-    # Relationships (one-directional; notes are always queried from the note side)
-    books: Mapped[list["Book"]] = relationship(secondary=note_books, lazy="selectin")
-    chapters: Mapped[list["Chapter"]] = relationship(secondary=note_chapters, lazy="selectin")
-    highlights: Mapped[list["Highlight"]] = relationship(secondary=note_highlights, lazy="selectin")
-    tags: Mapped[list["Tag"]] = relationship(secondary=note_tags, lazy="selectin")
-
-    def __repr__(self) -> str:
-        """String representation of Note."""
-        return f"<Note(id={self.id}, title='{self.title}', kind={self.kind})>"
+    "reading_session_highlights",
+]
